refactor(executor): extract subshell and compound command logic into separate module
